requirements.txt
hopsworks==4.2.*
pandas<2.2.0
requests
scikit-learn
python-dotenv
pyarrow
confluent-kafka
hopsworks[python]python-dotenv
